+++ crates/tau-runtime/src/lib.rs
@@ -3,6 +3,7 @@ pub mod observability_loggers_runtime;
 pub mod rpc_capabilities_runtime;
 pub mod rpc_protocol_runtime;
 pub mod runtime_output_runtime;
+pub mod transport_conformance_runtime;
 pub mod transport_health;
 
 pub use channel_store::*;
@@ -10,4 +11,5 @@ pub use observability_loggers_runtime::*;
 pub use rpc_capabilities_runtime::*;
 pub use rpc_protocol_runtime::*;
 pub use runtime_output_runtime::*;
+pub use transport_conformance_runtime::*;
 pub use transport_health::*;
